rename: renaming package adminAlarm to admin.alarm
